+++ R/checkPrep.R
@@ -10,7 +10,7 @@ checkPrep <- function(mydata, Names, type, remove.calm = TRUE, remove.neg = TRUE
   ## pre-defined types
   ## existing conditioning variables that only depend on date (which is checked)
   conds <- c(
-    "default", "year", "hour", "month", "season", "weekday",
+    "default", "year", "hour", "month", "season", "weekday", "week",
     "weekend", "monthyear", "gmtbst", "bstgmt", "dst", "daylight",
     "yearseason", "seasonyear"
   )