+++ NEWS.md
@@ -1,9 +1,12 @@
 ## openair 2.7-99
 
+## openair 2.7-6
+
 - fix bug with `percentileRose` when `statistic = "cpf"` and multiple pollutants
 - fix bug in `timeAverage` when `type = "season"` and `avg.time = "season"`
 - add `week` to openair default types
 - fix bug in `timeVariation` when considering difference plots with missing data
+- fix CRAN check when using `return`
 
 ## openair 2.7-4